requirements.txt
# ===============================================
# Trading Bot Dependencies - MT5 Version
# Author: BLESSING OMOREGIE
# GitHub: Nixiestone
# Install all dependencies with: pip install -r requirements.txt
# ===============================================

# CRITICAL: Windows-specific (for MT5 integration) - INSTALL FIRST
pywin32

# MetaTrader 5 Integration - CRITICAL FOR MT5
MetaTrader5

# Core Dependencies
python-dotenv
pydantic
pydantic-settings

# Data Processing and Analysis
numpy
pandas
scipy

# Technical Analysis
ta

# Machine Learning
scikit-learn
xgboost
tensorflow
keras
joblib

# Alternative Exchange Support (Optional)
ccxt
python-binance

# News Scraping and Sentiment
newspaper3k
beautifulsoup4
lxml
requests
textblob
vaderSentiment
transformers
torch
sentencepiece

# Database
sqlalchemy
alembic

# Notifications
python-telegram-bot
discord-webhook
discord.py

# Security and Encryption
cryptography
bcrypt
python-jose

# API and Web Framework
fastapi
uvicorn
aiohttp
websockets

# Monitoring and Metrics
prometheus-client
psutil

# Utilities
pytz
python-dateutil
schedule
retrying
tenacity

# Logging and Error Tracking
colorlog

# Testing
pytest
pytest-asyncio
pytest-cov
pytest-mock

# Code Quality
black
flake8

# Development Tools
ipython

# Async Support
asyncio
aiofiles

# HTTP and Networking
httpx
urllib3

# JSON and Data Serialization
orjson

# Environment and Configuration
configparser
click

# Rate Limiting
ratelimit

# Data Validation
cerberus

# Caching
cachetools

# Time Series Analysis
statsmodels

# Plotting (for analysis)
matplotlib
plotly
# Type Hints
typing-extensions

# Windows-specific (for MT5 integration)
pywin32

cryptography==41.0.7
python-telegram-bot==20.7
SQLAlchemy==2.0.23
alembic==1.13.0
redis==5.0.1
psycopg2-binary==2.9.9